feat: support frontier backend in lazy-loading mode
