+++ Cargo.toml
@@ -275,6 +275,7 @@ fc-rpc = { git = "https://github.com/moonbeam-foundation/frontier", branch = "mo
     "rpc-binary-search-estimate",
 ] }
 fc-rpc-core = { git = "https://github.com/moonbeam-foundation/frontier", branch = "moonbeam-polkadot-stable2412" }
+fc-rpc-v2-api = { git = "https://github.com/moonbeam-foundation/frontier", branch = "moonbeam-polkadot-stable2412" }
 fc-storage = { git = "https://github.com/moonbeam-foundation/frontier", branch = "moonbeam-polkadot-stable2412" }
 
 # Cumulus (wasm)
